Share settings link across mobile apps
