+++ mobile/apps/photos/lib/ui/settings/support/help_support_page.dart
@@ -78,12 +78,7 @@ class HelpSupportPage extends StatelessWidget {
             },
           ),
         ],
-        _SupportLink(
-          label: l10n.exportLogs,
-          onTap: () async {
-            await _exportLogs(context);
-          },
-        ),
+        SettingsLink(label: l10n.exportLogs, onTap: () => _exportLogs(context)),
         const SizedBox(height: 24),
         _sectionTitle(context, l10n.browseHelpPages),
         _buildHelpTopicItem(
@@ -125,11 +120,9 @@ class HelpSupportPage extends StatelessWidget {
           icon: HugeIcons.strokeRoundedWrench01,
           faqUrl: _troubleshootingFaqUrl,
         ),
-        _SupportLink(
+        SettingsLink(
           label: l10n.viewAllHelpTopics,
-          onTap: () async {
-            await _openHelpPage(context, title: l10n.helpAndSupport);
-          },
+          onTap: () => _openHelpPage(context, title: l10n.helpAndSupport),
         ),
         const SizedBox(height: 24),
       ],
@@ -202,40 +195,5 @@ class HelpSupportPage extends StatelessWidget {
         showShortToast(context, context.strings.somethingWentWrong);
       }
     }
-  }
-}
-
-class _SupportLink extends StatelessWidget {
-  final String label;
-  final Future<void> Function() onTap;
-
-  const _SupportLink({required this.label, required this.onTap});
-
-  @override
-  Widget build(BuildContext context) {
-    final colors = context.componentColors;
-    return GestureDetector(
-      behavior: HitTestBehavior.opaque,
-      onTap: onTap,
-      child: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 12),
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Text(
-              label,
-              style: TextStyles.bodyBold.copyWith(color: colors.primary),
-            ),
-            const SizedBox(width: 4),
-            HugeIcon(
-              icon: HugeIcons.strokeRoundedArrowRight01,
-              color: colors.primary,
-              size: 16,
-              strokeWidth: 1.6,
-            ),
-          ],
-        ),
-      ),
-    );
   }
 }
