+++ mobile/packages/ente_components/lib/ente_components.dart
@@ -19,6 +19,7 @@ export 'components/selection_controls/radio_component.dart';
 export 'components/selection_controls/toggle_switch_component.dart';
 export 'components/selection_summary_chip_component.dart';
 export 'components/settings_item.dart';
+export 'components/settings_link.dart';
 export 'components/settings_page_scaffold.dart';
 export 'components/slider_component.dart';
 export 'components/stepper_component.dart';
